+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add vendor directory to warble config for glimmer package command.
 - Make WidgetProxy register only the nearest ancestor property observer, calling on_modify_text and on_widget_selected for widgets that support these listeners, or otherwise the widget specific customizations
+- Add glimmer package:clean command
 
 ## 0.2.2